Revert "fixed minor issues"
