--- Assets/Materials/EtherealLineMaterial.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f1bf65afc4d4b644cae319664f330df3
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
