--- Assets/Materials/EtherealParticleMaterial.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 00e0af8055a4c45439f2a5c3d6788e14
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
