--- Assets/Materials/GroundIndicatorMaterial.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9b60929b71273cd49b0b37e3a9b2e8cb
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
