--- Assets/Mechanics.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 14456726370df7d498b5981183f8f961
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
