--- Assets/Mechanics/Abilities.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cdda300985c1a854b9c86804e4fd7bb6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
