--- Assets/Mechanics/Abilities/DistortionAnimator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1371d67316c0640479669004a34389c2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
