--- Assets/Mechanics/Abilities/WarpMarker.prefab
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:051eb2b76b76d5ac3c78290f1037168f324d230a213db057e0f4d119c9c6ab89
-size 125213
