+++ Assets/ModularFirstPersonController/FirstPersonController/FirstPersonController.prefab
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:6203ddca6db362527c0880e0d63b8acc3cabe8961c7a48e0db49761d318fe0d2
-size 24622
+oid sha256:8b66dbf65749cb3926a9c6be022b2d7a6ed5f36a7b947f566a1ceff83c3cf9d2
+size 20240
