+++ Assets/ModularFirstPersonController/FirstPersonController/WarpEffect.prefab
