--- Assets/Polytope Studio/Lowpoly_Environments/URP/PT_Nature_Free_URP_12.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 194ffb9f0c5b86f4aaa6ac1ef3ade1ed
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
