--- Assets/Polytope Studio/Lowpoly_Environments/URP/PT_Nature_Free_URP_14.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 713ea1b84c2a9fc4d8c7f695ca916c8b
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
