--- Assets/Polytope Studio/Lowpoly_Environments/URP/PT_Nature_Free_URP_16.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ac56f8aebac30674489c8b5741b0489a
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
