--- Assets/Polytope Studio/Lowpoly_Environments/URP/PT_Nature_Free_URP_17.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 68d9417a84829c6459075121b8ac78bd
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
