+++ Assets/Scenes/Testing/TestLevelOne.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:a7598b454a367386d9edf44e3a55c688f76c44f3f78f3406fbe534ffcfe137af
-size 55858
+oid sha256:758be05c2f3bfc5d4752344e298464270f163bb9e92dad5373435570cd57a60e
+size 52821
