--- Assets/Scenes/Testing/ui-warpCooldown.unity
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:fb89f1023620ea8eab2872ad90d0a96c925d7508bac88c0f1ab1ddac9f9dadd9
-size 14344
