--- Assets/Scenes/Testing/ui-warpCooldown.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ec0ee94b3c34c5648a39d250f824b440
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
