--- Assets/UI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ff50c6e278cb7994e8eaa0834d084a32
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
