--- Assets/UI/CooldownCanvas.prefab
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:78bfb06fdb504b7d4f8419366e2a25ec3a4eb4231e286423f82db85a82b62ac9
-size 7347
