--- Assets/UI/CooldownCanvas.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3564bb89ceaff8a4bb7ef2c349a82bd9
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
