--- Assets/WarpOverlay.prefab
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:07809f59fdcf5c4de43896f002e0af88b614783484aeb329e3ca6f22744855a9
-size 2198
