--- Assets/WarpOverlay.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 363831b356d44ca45a222b5bbd38e944
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
